\name{lpjgr-package}
\alias{lpjgr-package}
\alias{lpjgr}
\docType{package}
\title{
  R Wrapper around some LPJ-Guess functions.
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Drew Holzworth <d.holzworth@westernsydney.edu.au>

Maintainer: Drew Holzworth <d.holzworth@westernsydney.edu.au>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
